webapi/webapi-appuri-w3c-tests/appuri/URI.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset="utf-8">
<title>URI scheme Test: App URI</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/app-uri/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  String.prototype.EndWith = function (s) {
    if(s == null || s == "" || this.length == 0 || s.length > this.length)
      return false;
    if(this.substring(this.length - s.length) == s)
      return true;
    else
      return false;
    return true;
  }

  String.prototype.StartWith = function (s) {
    if(s == null || s == ""|| this.length == 0 || s.length > this.length)
      return false;
    if(this.substr(0, s.length) == s)
      return true;
    else
      return false;
    return true;
  }

  var loc;
  setup(function () {loc =  window.location;});

  test(function () {
    assert_equals(loc.protocol, "app:", "protocol of window.location");
  }, "Check if the protocol value of window.location is right");

  test(function () {
    assert_true(loc.href.StartWith("app://"), "href of window.location is start with app://");
    assert_true(loc.href.EndWith("/opt/webapi-appuri-w3c-tests/appuri/URI.html"), "href of window.location is end with /opt/webapi-appuri-w3c-tests/appuri/URI.html"+loc.href);
  }, "Check if the href value of window.location is right");

  test(function () {
    assert_true(loc.origin.StartWith("app://"), "origin of window.location is start with app://");
    assert_not_equals(loc.origin, "app://", "origin of window.location");
  }, "Check if the origin value of window.location is right");

  test(function () {
    assert_equals(loc.pathname, "/opt/webapi-appuri-w3c-tests/appuri/URI.html", "pathname of window.location");
  }, "Check if the pathname value of window.location is right");

  test(function () {
    assert_equals(loc.hash, "", "hash of window.location");
  }, "Check if the hash value of window.location is right");

  test(function () {
    assert_equals(loc.port, "", "port of window.location");
  }, "Check if the port value of window.location is right");

  test(function () {
    var img = document.createElement("img");
    img.src = "image/test.png";
    assert_true(img.src.StartWith("app://"), "src of image is start with app://");
    assert_true(img.src.EndWith("/opt/webapi-appuri-w3c-tests/appuri/image/test.png"), "src of image is end with /opt/webapi-appuri-w3c-tests/appuri/image/test.png");
  }, "Check if the src value of image is right");
</script>
